crosvm: linux: device_helper: add VirtioDeviceBuilder trait
As we are going to jail vhost-user devices, we will need a way to obtain
a Minijail that we can use in that case. This CL introduces a
VirtioDeviceBuilder trait that can be implemented for any device
configuration type and allows to spawn a configured device as well as
its jail, which is created from the correct policy file depending on the
virtio transport desired. In effect, it splits the creation of the
device and jail into two separate methods so the same jail creation code
can be reused with both regular virtio and vhost-user.

In order to illustrate, the console device is converted to this new
scheme. Other devices will follow when relevant.

BUG=b:217480043
TEST=serial device works with `crosvm run`.

Change-Id: I6b59d3794580bbdabce6df290c9ab8262d8e178c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3706491
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>