devices: vvu: update vvu to frame full messages
Update the vvu driver so that it sends and receives data as discrete
vhost-user messages instead of as a byte stream. This will make it
easier for vvu to intercept transmitted messages, and it allows the
receiving code to be somewhat simplified.

The device implementation already processes received data as full
messages, and doesn't currently examine transmitted messages, so it
doesn't need to be updated.

BUG=b:201745804
TEST=launch sibling VM on manatee

Change-Id: I94a7437a28136e682d2002a34e5f0b7d29d9c238
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3765015
Tested-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>