virtio: vhost-user: console: add type to reset stdin to canon mode
Add a type that, when dropped, resets the standard input to canon mode
if needed. We were doing it manually so far, which is error-prone if we
exit early due to an error.

This new type is also required for properly implementing the new
`devices` command.

BUG=b:217480043
TEST=vhost-user console device works on Linux.

Change-Id: I36d55f248198dc6383ae8e5557ac394baaab6e70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3725736
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>