devices: iommu: introduce iommu faults
The current virtio-iommu implementation presents an IOVA-to-GPA
translation API to virtio devices. However, this API is insufficient for
proper protection because it does not allow the virtio-iommu device to
revoke translations if the guest kernel removes a mapping that was used
for a previous translation request.

The fact that VVU devices are controlled via VFIO in ManaTEE means the
current translation-based implementation is especially problematic. If a
VVU device process crashes, the guest kernel will free any memory the
process mapped into the VVU proxy via VFIO. However, this is not
communicated to the VVU proxy, so the VVU proxy may continue to write to
this freed memory. Now that VVU supports mapping GPU buffers into
sibling VMs, this can also expose random CrOS guest memory to sibling
VMs. Furthermore, this use-after-free issue can also potentially be
taken advantage of by a malicious userspace program.

To address these shortcomings, this change switches the API presented by
virtio-iommu to other virtio devices from a translation based API to an
export based API. When using this API, a virtio device (e.g. VVU) first
starts an export session. From there, the virtio device exports memory
from the virtio-iommu device when it needs to access the memory and
releases the exported memory when it is done with the memory. If the
guest unmaps memory that is currently exported, then the virtio-iommu
device generates an "iommu fault". This is communicated to the other
virtio device by signaling an eventfd returned when the export session
was started. After an iommu fault occurs, the target virtio device must
immediately release all exported memory. Furthermore, the current export
session is dead, so a new one must be started before more memory can be
exported.

The virtio-iommu device prevents memory that is exported from being
reused within the guest by not completing the unmap (or detach)
operation until the generated iommu fault is resolved. This is
accomplished via a second eventfd returned from the MemoryMapper to the
virtio-iommu device when an iommu fault occurs. The virtio-iommu device
then waits on that event until the iommu fault is resolved.

On the VVU side of things, an export session is scoped to the lifespan
of the worker thread. For VVU, there are two types of exported memory.
The first is the memory associated with the virtqueues. This is managed
via the new ExportedRegion struct. The second is memory that is mapped
into the sibling VM via SHMEM_MAP. When an iommu fault occurs, the VVU
worker thread sends SHMEM_UNMAP requests for all mapped memory and then
exits, which automatically drops and releases all ExportedRegion
structs.

BUG=b:237620529
TEST=boot manatee sibling, kill cros device, observe that the
corresponding vvu proxy exits before virtio-iommu detaches the device.

Change-Id: I0ceb822dbba7269a74cea814b93e8184a102b2fa
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3765081
Commit-Queue: David Stevens <stevensd@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>