devices: iommu: replace memory_mapper::Error with anyhow
Replace thiserror enum with anyhow. The only error that needs to be
checked for is IovaPartialOverlap, which can be handled by changing the
add/remove map return types from () to bool. This CL also adds proper
handling of that case to the iommu device.

BUG=b:237620529
TEST=compiles

Change-Id: I066cf2d609d1264b235b019ec098f9a761535f9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3737405
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>