devices: vfio: report preferred PCI address
Rather than deferring the choice of PCI address to runtime during the
allocate_address() function, move it to the constructor so an error
message can be returned earlier. This lets us report the preferred
address in the PciDevice impl so that the PCI address allocation code in
arch can reserve the preferred address before allocating addresses for
other devices.

BUG=b:237415650
TEST=tools/presubmit --all

Change-Id: I8a00f5f4251450f475bc83c346b3288dbf901f1d
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3786973
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>