virtio: vhost-user: block: leverage BlockAsync for device creation
The vhost-user BlockBackend is clearly based on the regular virtio
BlockAsync, as is attested by the copy/pasted code between the
constructors of both implementations. Leverage the new VhostUserDevice
trait to build a BlockBackend from a BlockAsync. On top of removing
quite a bit of duplicated code and making the disk ID feature available
to vhost-user, this will also allow us to spawn block devices using the
new devices command.

BlockAsync has members to manage runtime that we don't use for
vhost-user, but they remain inactive as its lifetime is very limited in
our use-case. And the de-duplication of code largely balances this small
inelegancy.

BUG=b:217480043
TEST=`crosvm device block` creates a block device that is usable.

Change-Id: I4a51400f26cec286c7ba2f4f3cd1d7ad75594972
Reviewed-on: https://chromium-review.googlesource.com/c/crosvm/crosvm/+/3762976
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>